route_error refactored
