+++ server/src/main.rs
@@ -1,51 +1,21 @@
-use rocket::{Build, Rocket, launch, routes, catchers};
+use rocket::{Build, Rocket, launch, routes, catchers, get};
 use std::env;
 use rocket::tokio;
 use tokio_postgres::NoTls;
 use std::net::Ipv4Addr;
+use rocket::serde::json::Json;
 
 mod api;
 mod db;
 mod tools;
 mod qr;
 
-mod route_tools {
-    use rocket::{get, catch, Responder};
-    use rocket::serde::json::Json;
+use tools::route_error;
 
-    #[get("/")]
-    pub fn ping() -> Json<&'static str> {
-        Json("Eatup up and running!")
-    }
-
-    #[catch(404)]
-    pub fn qr_not_found() -> Json<&'static str> {
-        Json("QR not found. Are you sure the QR should be valid?")
-    }
-
-    #[catch(501)]
-    pub fn api_not_implemented() -> Json<&'static str> {
-        Json("Ups, this feature is not implemented yet.")
-    }
-
-    #[catch(500)]
-    pub fn internal_server_error() -> Json<&'static str> {
-        Json("Well, this is embarrassing. Something went wrong on our side. Turns out, rust can fail too.")
-    }
-
-    #[derive(Responder)]
-    #[response(status = 409, content_type = "json")]
-    pub struct InvalidAPI {
-        message: String
-    }
-
-    impl InvalidAPI {
-        pub fn new(message: String) -> InvalidAPI {
-            InvalidAPI { message }
-        }
-    }
+#[get("/")]
+pub fn ping() -> Json<&'static str> {
+    Json("Eatup up and running!")
 }
-
 
 #[launch]
 async fn rocket() -> Rocket<Build> {
@@ -83,9 +53,9 @@ async fn rocket() -> Rocket<Build> {
     };
     rocket::custom(&config)
         .manage(client)
-        .mount("/", routes![route_tools::ping])
+        .mount("/", routes![ping])
         .mount("/api/v1", api::get_v1_routes())
-        .register("/api", catchers![route_tools::api_not_implemented])
+        .register("/api", catchers![route_error::api_not_implemented])
         .mount("/", rocket::fs::FileServer::from("/db/public"))
-        .register("/qr/", catchers![route_tools::qr_not_found, route_tools::internal_server_error])
+        .register("/qr/", catchers![route_error::qr_not_found, route_error::internal_server_error])
 }
