+++ server/src/tools/mod.rs
@@ -1,3 +1,4 @@
+pub mod route_error;
 mod uuidwrapper;
 
 pub use uuidwrapper::UuidWrapper;
